Added HScroll to module import log. Echo "Failed" messages to warning window. Pause before closing import window. Fixes #51
